게시판 백엔드작업

+++ EcoInsight/src/main/java/com/semi/ecoinsight/community/model/dto/CommunityDTO.java
@@ -21,7 +21,7 @@ public class CommunityDTO {
 	private String cBoardContent;
 	private Date createDate;
 	private Date modifiedDate;
-	private Long viewCount;
+	private int viewCount;
 	private String isDeleted;
 
 }
